[flutter] Move player implementations into separate files
